+++ packages/core/src/model-types.ts
@@ -53,8 +53,27 @@ export type AwsBedrockChatModel = LiteralUnion<
 >;
 
 /**
+ * AWS Bedrock models for embedding.
+ */
+export type AwsBedrockEmbeddingModel = LiteralUnion<'amazon--titan-embed-text'>;
+
+/**
+ * Perplexity models for chat completion.
+ */
+export type PerplexityChatModel = LiteralUnion<'sonar' | 'sonar-pro'>;
+
+/**
  * AI Core open source models for chat completion.
  */
 export type AiCoreOpenSourceChatModel = LiteralUnion<
-  'mistralai--mistral-large-instruct' | 'mistralai--mistral-small-instruct'
+  | 'cohere--command-a-reasoning'
+  | 'mistralai--mistral-large-instruct'
+  | 'mistralai--mistral-medium-instruct'
+  | 'mistralai--mistral-small-instruct'
 >;
+
+/**
+ * AI Core sap managed model for embedding.
+ */
+export type AiCoreOpenSourceEmbeddingModel =
+  LiteralUnion<'nvidia--llama-3.2-nv-embedqa-1b'>;