+++ packages/document-grounding/src/client/api/schema/sdm-pipeline-create-request.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2025 SAP SE or an SAP affiliate company. All rights reserved.
+ *
+ * This is a generated file powered by the SAP Cloud SDK for JavaScript.
+ */
+import type { MetaData } from './meta-data.js';
+/**
+ * Representation of the 'SDMPipelineCreateRequest' schema.
+ */
+export type SDMPipelineCreateRequest = {
+  type: 'SDM';
+  metadata: MetaData;
+} & Record<string, any>;