+++ packages/document-grounding/src/client/api/schema/search-pipelines-response.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2025 SAP SE or an SAP affiliate company. All rights reserved.
+ *
+ * This is a generated file powered by the SAP Cloud SDK for JavaScript.
+ */
+import type { PipelineMinimalResponse } from './pipeline-minimal-response.js';
+/**
+ * Representation of the 'SearchPipelinesResponse' schema.
+ */
+export type SearchPipelinesResponse = {
+  count?: number;
+  resources: PipelineMinimalResponse[];
+} & Record<string, any>;