+++ packages/orchestration/src/util/translation.test.ts
@@ -1,14 +1,34 @@
 import { buildTranslationConfig } from './translation.js';
-import type { SAPDocumentTranslation } from '../client/api/schema/index.js';
+import type {
+  SAPDocumentTranslationInput,
+  SAPDocumentTranslationOutput
+} from '../client/api/schema/index.js';
 
 describe('Translation module config', () => {
-  it('builds translation config', async () => {
+  it('builds input translation config', async () => {
     const translationConfig = buildTranslationConfig('input', {
       sourceLanguage: 'de-DE',
       targetLanguage: 'en-US'
     });
 
-    const expectedTranslationConfig: SAPDocumentTranslation = {
+    const expectedTranslationConfig: SAPDocumentTranslationInput = {
+      type: 'sap_document_translation',
+      config: {
+        source_language: 'de-DE',
+        target_language: 'en-US'
+      }
+    };
+
+    expect(translationConfig).toEqual(expectedTranslationConfig);
+  });
+
+  it('builds output translation config', async () => {
+    const translationConfig = buildTranslationConfig('output', {
+      sourceLanguage: 'de-DE',
+      targetLanguage: 'en-US'
+    });
+
+    const expectedTranslationConfig: SAPDocumentTranslationOutput = {
       type: 'sap_document_translation',
       config: {
         source_language: 'de-DE',