+++ packages/orchestration/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @sap-ai-sdk/orchestration
 
+## 2.3.0
+
+### Minor Changes
+
+- 7c9605d: [Compatibility Note] `zod` was upgraded to v4
+- 485e21b: [New Functionality] Added support for `protected_material_code` property to `buildAzureContentSafetyFilter()` function for output filter configuration to allow detecting protected code content from known github repositories.
+
+### Patch Changes
+
+- Updated dependencies [7c9605d]
+  - @sap-ai-sdk/prompt-registry@2.3.0
+  - @sap-ai-sdk/ai-api@2.3.0
+  - @sap-ai-sdk/core@2.3.0
+
+## 2.2.0
+
+### Minor Changes
+
+- 6100bca: [Improvement] Add `cohere--command-a-reasoning`, `mistralai--mistral-medium-instruct` and perplexity-ai `sonar` and `sonar-pro` to model list
+- 347eac1: [New Functionality] Introduce orchestration embedding client for consuming embedding feature of the orchestration service.
+
+### Patch Changes
+
+- Updated dependencies [6100bca]
+- Updated dependencies [5225275]
+- Updated dependencies [347eac1]
+- Updated dependencies [58464e9]
+  - @sap-ai-sdk/core@2.2.0
+  - @sap-ai-sdk/ai-api@2.2.0
+  - @sap-ai-sdk/prompt-registry@2.2.0
+
 ## 2.1.0
 
 ### Patch Changes