+++ sample-cap/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sap-ai-sdk/sample-cap
 
+## 2.3.0
+
+### Patch Changes
+
+- Updated dependencies [7c9605d]
+- Updated dependencies [485e21b]
+  - @sap-ai-sdk/orchestration@2.3.0
+  - @sap-ai-sdk/ai-api@2.3.0
+  - @sap-ai-sdk/foundation-models@2.3.0
+
+## 2.2.0
+
+### Patch Changes
+
+- Updated dependencies [6100bca]
+- Updated dependencies [347eac1]
+- Updated dependencies [58464e9]
+  - @sap-ai-sdk/orchestration@2.2.0
+  - @sap-ai-sdk/ai-api@2.2.0
+  - @sap-ai-sdk/foundation-models@2.2.0
+
 ## 2.1.0
 
 ### Patch Changes